pyproject.toml
[project]
name = "tana-auto-tagger"
version = "0.1.0"
description = "Auto-tag untagged Tana notes using local AI classification"
requires-python = ">=3.10"
dependencies = [
    "sentence-transformers>=2.2.0",
    "httpx>=0.25.0",
    "typer[all]>=0.9.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "python-telegram-bot[webhooks]>=20.0",
    "dateparser>=1.2.0",
]

[project.scripts]
tana-tagger = "tana_auto_tagger.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
